fix(constant): move RenderFileType to constant, fix STL exports flag
seems prusaslicer (or openscad) has issues with openscad bin stl exports. utilize ascii for now.
